Move logger test fixtures to a separate file
